==> 02_Data_Cleaning_EDA.ipynb <==
import pandas as pd
import pickle


Load results df - that is first attempt api response

results_df = pd.read_pickle("./results_df.pkl")



##### Load all categories pkl files from API response

with open('bracelet.pkl', 'rb') as f:
    data_list_bracelet = pickle.load(f)
    
with open('earrings.pkl', 'rb') as f:
    data_list_earrings = pickle.load(f)
    
with open('necklace.pkl', 'rb') as f:
    data_list_necklace = pickle.load(f)

with open('anklet.pkl', 'rb') as f:
    data_list_anklet = pickle.load(f)

with open('brooch.pkl', 'rb') as f:
    data_list_brooch = pickle.load(f)

with open('ring.pkl', 'rb') as f:
    data_list_ring = pickle.load(f)

with open('pendant.pkl', 'rb') as f:
    data_list_pendant = pickle.load(f)
    
with open('piercing.pkl', 'rb') as f:
    data_list_piercing = pickle.load(f)
    

##### Convert to df

jewelry_df_bracelet = pd.DataFrame(data_list_bracelet)
jewelry_df_earrings = pd.DataFrame(data_list_earrings)
jewelry_df_necklace = pd.DataFrame(data_list_necklace)
jewelry_df_anklet = pd.DataFrame(data_list_anklet)
jewelry_df_brooch = pd.DataFrame(data_list_brooch)
jewelry_df_ring = pd.DataFrame(data_list_ring)
jewelry_df_pendant = pd.DataFrame(data_list_pendant)
jewelry_df_pierching = pd.DataFrame(data_list_piercing)


##### EDA & Cleaning

print('Bracelet',jewelry_df_bracelet.listing_id.nunique())
print('Earrings',jewelry_df_earrings.listing_id.nunique())
print('Necklace',jewelry_df_necklace.listing_id.nunique())
print('Anklet',jewelry_df_anklet.listing_id.nunique())
print('Brooch',jewelry_df_brooch.listing_id.nunique())
print('Ring',jewelry_df_ring.listing_id.nunique())
print('Pendant',jewelry_df_pendant.listing_id.nunique())
print('Piercing',jewelry_df_pierching.listing_id.nunique())


jewelry_df_ring.category_path[1]


', '.join(jewelry_df_ring.category_path[1])



jewelry_df_ring.head()



##### Merge all sub dfs to one jewelry df

jewelry_df_bracelet_clean = jewelry_df_bracelet.copy()
# jewelry_df_earrings
# jewelry_df_necklace
# jewelry_df_anklet
# jewelry_df_brooch
# jewelry_df_ring
# jewelry_df_pendant
# jewelry_df_pierching


# jewelry_df_bracelet_clean.head()

#columns interesting for clustering - description, featured_rank, has_var?,customizable, materials,
#num_fav,ocassion


#Pickle Cleaned DF ---
jewelry_df_bracelet_clean.to_pickle('clean_bracelet_df.pkl')

#Columns that can be dropped - 'category_id','category_path_ids','shipping_template_id',
                                #'shop_section_id','suggested_taxonomy_id','taxonomy_id','url'

#CONVERT PRICE TO USD

# jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop_duplicates(keep = 'first')

# jewelry_df_bracelet_clean.head()


def convert_list_col(lst):
    try:
        return ', '.join(lst)
    except TypeError:
        return ''


convert_list_col(jewelry_df_bracelet_clean.category_path[1])

jewelry_df_bracelet_clean.category_path = jewelry_df_bracelet_clean.category_path.apply(convert_list_col)

jewelry_df_bracelet_clean.category_path.value_counts()

#Drop columns from jewelry clean bracelet df -- 'category_id','category_path_ids','shipping_template_id',
                                #'shop_section_id','suggested_taxonomy_id','taxonomy_id','url'

    
jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['category_id','category_path_ids',
                                                                    'shipping_template_id',
                                'shop_section_id','suggested_taxonomy_id','taxonomy_id','url'])

jewelry_df_bracelet_clean.head()

#### Featured Rank Cleaning

Featured rank is assigned to a product when its on the featured listing -- closer to 0 higher the rank. Setting all nans to a large number (something like 9999 to portray that is isnt featured)

# jewelry_df_bracelet_clean[~jewelry_df_bracelet_clean.featured_rank.isna()]
# jewelry_df_bracelet_clean.featured_rank.min()

jewelry_df_bracelet_clean.featured_rank = jewelry_df_bracelet_clean.featured_rank.fillna(9999)

#### Boolean Fields convert

is_customizable - convert Nan to 0s and convert the column to 0(False) and 1(True)

# jewelry_df_bracelet_clean[jewelry_df_bracelet_clean.is_supply.isna()]
jewelry_df_bracelet_clean.is_supply.value_counts()

jewelry_df_bracelet_clean.is_customizable = jewelry_df_bracelet_clean.is_customizable.fillna(False)
jewelry_df_bracelet_clean.is_digital = jewelry_df_bracelet_clean.is_digital.fillna(False)
jewelry_df_bracelet_clean.is_supply = jewelry_df_bracelet_clean.is_supply.fillna(False)

def convert_bool_col(boolc):
    return int(boolc)

def convert_str_bool_col(strcol):
    if type(strcol) == str:
        if strcol == 'false':
            return 0
        elif strcol == 'true':
            return 1
    elif type(strcol) == bool:
        return int(strcol)


convert_str_bool_col(True)

jewelry_df_bracelet_clean.is_customizable = jewelry_df_bracelet_clean.is_customizable.apply(convert_bool_col)
jewelry_df_bracelet_clean.is_digital = jewelry_df_bracelet_clean.is_digital.apply(convert_bool_col)
jewelry_df_bracelet_clean.is_supply = jewelry_df_bracelet_clean.is_supply.apply(convert_str_bool_col)

####  Drop More Columns

jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['file_data'])
jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['has_variations'])
jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['is_private'])
jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['item_dimensions_unit','item_height',
                                                                   'item_length','item_weight','item_weight_unit',
                                                                    'item_width'])

#### Records that have text fields in languages other than English -- drop! (150 records for bracelets out of ~50k)

# jewelry_df_bracelet_clean[jewelry_df_bracelet_clean.language.isna()]

# jewelry_df_bracelet_clean.language = jewelry_df_bracelet_clean.language.fillna('na')

jewelry_df_bracelet_clean.language.value_counts()

jewelry_df_bracelet_clean = jewelry_df_bracelet_clean[jewelry_df_bracelet_clean.language.isin(['en-US','MACHINE_en'])]


#### Drop Date fields - creation, ending, last_modified date, original creation, state date

jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['creation_tsz','last_modified_tsz',
                                                                   'original_creation_tsz','state_tsz'])
jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['language'])

#### Convert Materials - list to string

jewelry_df_bracelet_clean.materials.value_counts()

def convert_list_col_notry(lst):
    return ', '.join(lst)

jewelry_df_bracelet_clean.materials = jewelry_df_bracelet_clean.materials.apply(convert_list_col_notry)

#### Ocassion & Descr

jewelry_df_bracelet_clean.occasion = jewelry_df_bracelet_clean.occasion.fillna('None')

jewelry_df_bracelet_clean.description = jewelry_df_bracelet_clean.description.fillna('None')

#### Processing min and max

jewelry_df_bracelet_clean.processing_max = jewelry_df_bracelet_clean.processing_max.fillna(0)
jewelry_df_bracelet_clean.processing_min = jewelry_df_bracelet_clean.processing_min.fillna(0)

#### Recipient

jewelry_df_bracelet_clean.recipient = jewelry_df_bracelet_clean.recipient.fillna('any')

jewelry_df_bracelet_clean.recipient.value_counts()

#### Should Auto Renew, SKU -- drop column not needed

# jewelry_df_bracelet_clean.should_auto_renew.value_counts()

# jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['should_auto_renew'])
# jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['sku'])
# jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['state'])
# jewelry_df_bracelet_clean = jewelry_df_bracelet_clean.drop(columns=['style'])

#### Tags - list to str

jewelry_df_bracelet_clean.tags = jewelry_df_bracelet_clean.tags.apply(convert_list_col_notry)

jewelry_df_bracelet_clean.tags.head()

jewelry_df_bracelet_clean.taxonomy_path = jewelry_df_bracelet_clean.taxonomy_path.apply(convert_list_col_notry)

jewelry_df_bracelet_clean.taxonomy_path.head()

jewelry_df_bracelet_clean.title = jewelry_df_bracelet_clean.title.fillna('None')

jewelry_df_bracelet_clean.used_manufacturer = jewelry_df_bracelet_clean.used_manufacturer.apply(convert_bool_col)

jewelry_df_bracelet_clean.used_manufacturer.head()

jewelry_df_bracelet_clean.who_made = jewelry_df_bracelet_clean.who_made.fillna('someone_else')

def when_made_catg(when_made):
    if when_made == '2010_2019' or when_made == '2000_2009':
        return 'standard'
    elif when_made == 'made_to_order':
        return when_made
    else:
        return 'vintage'

jewelry_df_bracelet_clean.when_made = jewelry_df_bracelet_clean.when_made.apply(when_made_catg)

jewelry_df_bracelet_clean.when_made.value_counts()

jewelry_df_bracelet_clean.isna().sum()

jewelry_df_bracelet_clean.head()

#conversion rate on 9/8 - 

def convert_curr(curr_code, price):
    curr_dict = {
    'USD':1,
    'EUR':1.11,
    'GBP':1.23,
    'CAD':0.76,
    'AUD':0.68,
    'ILS':0.28,
    'CHF':1.01,
    'HKD':0.13,
    'SEK':0.1,
    'SGD':0.72,
    'NZD':0.64,
    'MXN':0.05,
    'NOK':0.11,
    'DKK':0.15,
    'JPY':0.01,
    'THB':0.03,
    'HUF':0.003,
    'TWD':0.03,
    'PHP':0.02}
    
    if curr_code in curr_dict.keys():
        return price*curr_dict[curr_code]
    else:
        return price

float(jewelry_df_bracelet_clean.price[0])

jewelry_df_bracelet_clean['price_usd'] = \
jewelry_df_bracelet_clean.apply(lambda x: convert_curr(x['currency_code'],float(x['price'])), axis=1)

### Clean df function

def clean_df(jewelry_df):
    
    #Create Copy
    jewelry_df_clean = jewelry_df.copy()
    
    #Filter - remove all products where language != english
    jewelry_df_clean.language = jewelry_df_clean.language.fillna('na')
    jewelry_df_clean = jewelry_df_clean[jewelry_df_clean.language.isin(['en-US','MACHINE_en'])]
    
    
    #Drop Columns
    jewelry_df_clean = jewelry_df_clean.drop(columns=['category_id','category_path_ids','shipping_template_id',
                                                      'shop_section_id','suggested_taxonomy_id','taxonomy_id',
                                                      'url','file_data','has_variations','is_private','style',
                                                      'item_dimensions_unit','item_height','item_length',
                                                      'item_weight','item_weight_unit','item_width',
                                                      'creation_tsz','last_modified_tsz','original_creation_tsz',
                                                      'state_tsz','language','should_auto_renew','sku','state'])
    
    #Impute Missing Values
    jewelry_df_clean.featured_rank = jewelry_df_clean.featured_rank.fillna(9999)
    jewelry_df_clean.is_customizable = jewelry_df_clean.is_customizable.fillna(False)
    jewelry_df_clean.is_digital = jewelry_df_clean.is_digital.fillna(False)
    jewelry_df_clean.is_supply = jewelry_df_clean.is_supply.fillna(False)
    jewelry_df_clean.occasion = jewelry_df_clean.occasion.fillna('None')
    jewelry_df_clean.description = jewelry_df_clean.description.fillna('None')
    jewelry_df_clean.processing_max = jewelry_df_clean.processing_max.fillna(0)
    jewelry_df_clean.processing_min = jewelry_df_clean.processing_min.fillna(0)
    jewelry_df_clean.recipient = jewelry_df_clean.recipient.fillna('any')
    jewelry_df_clean.title = jewelry_df_clean.title.fillna('None')
    jewelry_df_clean.who_made = jewelry_df_clean.who_made.fillna('someone_else')
    
    
    #Clean Columns
    jewelry_df_clean.category_path = jewelry_df_clean.category_path.apply(convert_list_col)
    jewelry_df_clean.is_customizable = jewelry_df_clean.is_customizable.apply(convert_bool_col)
    jewelry_df_clean.is_digital = jewelry_df_clean.is_digital.apply(convert_bool_col)
    jewelry_df_clean.is_supply = jewelry_df_clean.is_supply.apply(convert_str_bool_col)
    jewelry_df_clean.materials = jewelry_df_clean.materials.apply(convert_list_col_notry)
    jewelry_df_clean.tags = jewelry_df_clean.tags.apply(convert_list_col_notry)
    jewelry_df_clean.taxonomy_path = jewelry_df_clean.taxonomy_path.apply(convert_list_col)
    jewelry_df_clean.used_manufacturer = jewelry_df_clean.used_manufacturer.apply(convert_bool_col)
    jewelry_df_clean.when_made = jewelry_df_clean.when_made.apply(when_made_catg)
    jewelry_df_clean['price_usd'] = \
    jewelry_df_clean.apply(lambda x: convert_curr(x['currency_code'],float(x['price'])), axis=1)
    
    #Drop currency code and old price columns
    jewelry_df_clean = jewelry_df_clean.drop(columns = ['price','currency_code'])
    
    return jewelry_df_clean

#### Create copies and run clean df on all categories for jewelry dfs

jewelry_df_earrings_clean = jewelry_df_earrings.copy()
jewelry_df_necklace_clean = jewelry_df_necklace.copy()
jewelry_df_anklet_clean = jewelry_df_anklet.copy()
jewelry_df_brooch_clean = jewelry_df_brooch.copy()
jewelry_df_ring_clean = jewelry_df_ring.copy()
jewelry_df_pendant_clean = jewelry_df_pendant.copy()
jewelry_df_piercing_clean = jewelry_df_pierching.copy()

jewelry_df_piercing_clean2 = clean_df(jewelry_df_piercing_clean)

jewelry_df_anklet_clean2 = clean_df(jewelry_df_anklet_clean)
# jewelry_df_bracelet_clean2 = clean_df(jewelry_df_bracelet_clean)
jewelry_df_bracelet_clean2 = jewelry_df_bracelet_clean.copy()
jewelry_df_brooch_clean2 = clean_df(jewelry_df_brooch_clean)
jewelry_df_earrings_clean2 = clean_df(jewelry_df_earrings_clean)
jewelry_df_necklace_clean2 = clean_df(jewelry_df_necklace_clean)
jewelry_df_pendant_clean2 = clean_df(jewelry_df_pendant_clean)
jewelry_df_ring_clean2 = clean_df(jewelry_df_ring_clean)

jewelry_df_bracelet_clean2 = jewelry_df_bracelet_clean2.drop(columns = ['price','currency_code'])

#### Create identifier column and merge all dfs

jewelry_df_anklet_clean2['jewelry_catg'] = 'anklet'
jewelry_df_bracelet_clean2['jewelry_catg'] = 'bracelet'
jewelry_df_brooch_clean2['jewelry_catg'] = 'brooch'
jewelry_df_earrings_clean2['jewelry_catg'] = 'earrings'
jewelry_df_necklace_clean2['jewelry_catg'] = 'necklace'
jewelry_df_pendant_clean2['jewelry_catg'] = 'pendant'
jewelry_df_piercing_clean2['jewelry_catg'] = 'piercing'
jewelry_df_ring_clean2['jewelry_catg'] = 'ring'

jewelry_df = jewelry_df_anklet_clean2.append(jewelry_df_bracelet_clean2, ignore_index=True)

jewelry_df = jewelry_df.append(jewelry_df_brooch_clean2, ignore_index=True)

# jewelry_df = jewelry_df.append(jewelry_df_earrings_clean2, ignore_index=True)
# jewelry_df = jewelry_df.append(jewelry_df_necklace_clean2, ignore_index=True)
# jewelry_df = jewelry_df.append(jewelry_df_pendant_clean2, ignore_index=True)
# jewelry_df = jewelry_df.append(jewelry_df_piercing_clean2, ignore_index=True)
jewelry_df = jewelry_df.append(jewelry_df_ring_clean2, ignore_index=True)

jewelry_df.info()

jewelry_df.to_pickle('jewelry_df_clean.pkl')
