==> 04_Clustering_kmodes.ipynb <==
import pandas as pd
import numpy as np
from kmodes.kmodes import KModes
from kmodes.kprototypes import KPrototypes


#### Load Cleaned df from notebook#2

jewelry_df = pd.read_pickle('jewelry_df_clean.pkl')

jewelry_df.head()

jewelry_df.jewelry_catg.value_counts()

X = jewelry_df.copy()

#### Data Prep for clustering 
Drop columns - End Date, Description, Listing ID, Title, User ID, Featured Rank, Is Digital, Is Supply, Non Taxable, Recipient, Used Manufacturer

X = X.drop(columns=['description', 'ending_tsz', 'listing_id', 'title','user_id'])
X= X.drop(columns=['jewelry_catg'])

X.head()

#Drop more columns
#1. featured_rank
#2. is_digital
#3. is_supply
#4. non_taxable
#5. recipient
#6. used_manufacturer
X = X.drop(columns=['featured_rank','is_digital','is_supply','non_taxable','recipient','used_manufacturer'])

X.info()

#### Sample DataFrame to test clustering algorithm

X_test = X.sample(frac=0.2)

X_test = X_test.drop(columns=['category_path'])
X_test.info()

X_test.head()

X_test.to_pickle('test_df.pkl')

### K Prototype for clustering numeric + categorical fields 

(try 5 clusters)

kp = KPrototypes(n_clusters=5, init='Huang', n_init=1, verbose=True, n_jobs = -1)

kp.fit_predict(X_test, categorical = [0,1,3,7,8,10,11])

# print(kp.cluster_centroids_)
# print(kp.labels_)

pd.Series(kp.labels_).value_counts()
kp.

for elem in zip(kp.cluster_centroids_[0], kp.cluster_centroids_[1]):
    print(elem)
    print('******\n\n')
