==> 05_TextField_FeatureEng.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
from kmodes.kmodes import KModes
from kmodes.kprototypes import KPrototypes

test_df = pd.read_pickle('test_df.pkl')

test_df.head()

### One Hot Encode for Materials

#Silver - silver, sterling silver, 
#Gold
#Glass
#Diamond
#Stones - stone, gemstone, turquoise,quartz,amethyst,emrald,ruby
#pearl


test_df['contains_silver'] = test_df.materials.str.contains('silver', case = False, regex=False).apply(lambda x:1 if x else 0)
test_df['contains_gold'] = test_df.materials.str.contains('gold', case = False, regex=False).apply(lambda x:1 if x else 0)
test_df['contains_glass'] = test_df.materials.str.contains('glass', case = False, regex=False).apply(lambda x:1 if x else 0)
test_df['contains_diamond'] = test_df.materials.str.contains('diamond', case = False, regex=False).apply(lambda x:1 if x else 0)
test_df['contains_pearl'] = test_df.materials.str.contains('pearl', case = False, regex=False).apply(lambda x:1 if x else 0)
test_df['contains_stone'] = \
(test_df.materials.str.contains('quartz', case = False, regex=False) | \
 test_df.materials.str.contains('ruby', case = False, regex=False) | \
 test_df.materials.str.contains('emrald', case = False, regex=False) | \
 test_df.materials.str.contains('turquoise', case = False, regex=False) | \
 test_df.materials.str.contains('amethyst', case = False, regex=False) | \
 test_df.materials.str.contains('stone', case = False, regex=False) | \
 test_df.materials.str.contains('gemstone', case = False, regex=False)
).apply(lambda x:1 if x else 0)


#Can drop materials column now

test_df = test_df.drop(columns=['materials'])

test_df.info()

### One Hot Encode - Occasion

test_df.occasion.value_counts()

occasion_df = pd.get_dummies(test_df.occasion)

occasion_df = occasion_df.drop(columns=['None', 'baptism', 'bar_or_bat_mitzvah',
       'canada_day', 'chinese_new_year', 'cinco_de_mayo','confirmation', 'day_of_the_dead', 
        'easter', 'eid', 'fathers_day', 'get_well', 'graduation', 'halloween', 'hanukkah',
       'housewarming', 'july_4th', 'kwanzaa', 'new_baby',
       'new_years', 'prom', 'quinceanera', 'retirement', 'st_patricks_day',
       'sweet_16', 'sympathy', 'thanksgiving'])

# occasion_df.info()
# test_df_copy = test_df.copy()

test_df = pd.merge(test_df, occasion_df, left_index = True, right_index=True)


#can drop 'occasion' column

test_df = test_df.drop(columns=['occasion'])

### One Hot Encode - Tags Column ??? (DROP for now)

# test_df.tags.sample(20)
test_df = test_df.drop(columns=['tags'])

test_df.info()

### One Hot Encode - Taxonomy

# test_df.taxonomy_path.value_counts()

taxonomy_df = pd.DataFrame(test_df.taxonomy_path.copy())

taxonomy_df = taxonomy_df.taxonomy_path.str.replace(', ',',').str.get_dummies(sep=',')

taxonomy_df = taxonomy_df.drop(columns=['Art & Collectibles', 'Bands', 'Barbells',
       'Beaded Necklaces', 'Beads', 'Belly Rings','Bib Necklaces','Cabochons', 'Cameo Necklaces',
       'Chain & Link Bracelets', 'Chains', 'Charms', 'Charms & Pendants', 'Claddagh Rings',
       'Cluster Earrings', 'Coins & Money', 'Collectible Glass',
       'Craft Supplies & Tools','Ear Climbers', 'Ear Jackets', 'Ear Jackets & Climbers', 'Ear Weights',
       'Engagement Rings', 'Findings', 'Fraternal & Class Rings',
       'Friendship Bracelets', 'Gauge & Plug Earrings', 'Gems & Cabochons',
       'Hoop Earrings', 'ID & Medical Bracelets', 'Jewelry Sets', 'Keychains',
       'Keychains & Lanyards', 'Lanyards & Badge Holders',
       'Lariat & Y Necklaces', 'Lockets', 'Midi Rings','Collectibles','Clasps & Clips','Cuff & Wrap Earrings',
       'Monogram & Name Necklaces', 'Multi Strand Necklaces','Pendants',
       'Multistone Rings', 'Nipple Jewelry', 'Nose Rings & Studs',
       'Paper & Party Supplies', 'Party Favors', 'Party Favors & Games',
       'Party Supplies', 'Patterns & Blueprints', 'Patterns & How To',
       'Pet Collars & Leashes', 'Pet ID Tags', 'Pet Supplies', 'Pins & Clips',
       'Promise Rings', 'Ring Guards & Spacers', 'Screw Back Earrings',
       'Signet Rings', 'Stackable Rings', 'Statement Rings',
       'Tassel Necklaces', 'Threader Earrings',
       'Triplet & Double Rings', 'Tutorials', 'Wedding & Engagement','Accessories','Anniversary Rings',
        'Jewelry','Body Jewelry',
       'Wedding Bands', 'Woven & Braided Bracelets'])

taxonomy_df.sum().sort_values()

test_df = pd.merge(test_df, taxonomy_df, left_index = True, right_index=True)

#can drop taxonomy path

test_df = test_df.drop(columns=['taxonomy_path'])

test_df.info()

test_df.when_made.value_counts()

kp = KPrototypes(n_clusters=5, init='Huang', n_init=1, verbose=True, n_jobs = -1)

kp.fit_predict(test_df, categorical = [0,6,7,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34])

print(kp.cluster_centroids_)
# print(kp.labels_)

test_df_labeled = test_df.copy()

test_df_labeled['cluster_label'] = kp.labels_

test_df_labeled.cluster_label.value_counts()

pd.to_pickle(test_df_labeled,'test_labeled_df.pkl')

test_df.head()

plt.figure(figsize=(7,7))
plt.scatter(data_cat.x,data_cat.y, c = np.array(clusters), s = data_cat['size'], cmap= 'jet')
plt.title('clustered sensors - colors represent clusters', fontsize=14)
plt.xlim(-1,12), plt.ylim(-1,12)
plt.xlabel('x-coordinate', fontsize=14), plt.ylabel('y-coordinate', fontsize=14)
plt.show()
